requirements.txt
torch
transformers>=4.39.0
torchvision
pillow
datasets
accelerate
qwen_vl_utils

wandb

numpy
matplotlib
scikit-learn


## Qwen stuff (model access)
# qwen-vl or huggingface_hub
huggingface_hub
hf_xet
huggingface_hub[cli]


python-dotenv
datasets